wip on utility commands
